pathName through props and data in store folder
