+++ src/App.js
@@ -1,6 +1,6 @@
 import "./App.css";
 import Slider from "./components/Slider/Slider";
-import {Routes, Route} from "react-router-dom";
+import {Routes, Route, useLocation} from "react-router-dom";
 import Accordion from "./components/Accordion/Accordion";
 import Home from "./components/Home/Home";
 import Form from "./components/Form/Form";
@@ -11,8 +11,15 @@ import OrderList from "./components/OrderList/OrderList";
 import CardComponent from "./components/CardComponent/CardComponent";
 
 function App() {
+  // const location = useLocation();
+  // const currentPage = location.pathname;
+  // const filteredPathName = currentPage.replace('/', '');
+
+  // console.log("Current page : ",filteredPathName);
+
   return (
     <>
+      {/* <h1 className="sr-only hidden">{filteredPathName}</h1> */}
       <Routes>
         <Route path="/" element={<Hero />} />
         <Route path="/explore" element={<Home />} />
