+++ src/components/Accordion/Accordion.js
@@ -1,34 +1,8 @@
 import React, {useState} from "react";
 import className from "classnames";
+import DATA from '../../store/accordionData';
 
 const Accordion = () => {
-  // json data:
-  const DATA = [
-    {
-      fields: {
-        question: {
-          value: "Where is the sun located?",
-        },
-        answer: "Duis ac magna vel urna pharetra iaculis vel et nibh. Mauris facilisis venenatis tempor. Sed porta, ipsum dictum rhoncus porttitor, neque odio sodales elit, at tincidunt eros lacus nec felis. Cras a porttitor mi, at ullamcorper neque. Aliquam at leo bibendum, ornare arcu sit amet, maximus ante. Fusce id enim.",
-      },
-    },
-    {
-      fields: {
-        question: {
-          value: "Only human sustainable planet in the Solar System?",
-        },
-        answer: "Duis ac magna vel urna pharetra iaculis vel et nibh. Mauris facilisis venenatis tempor. Sed porta, ipsum dictum rhoncus porttitor, neque odio sodales elit, at tincidunt eros lacus nec felis. Cras a porttitor mi, at ullamcorper neque. Aliquam at leo bibendum, ornare arcu sit amet, maximus ante. Fusce id enim.",
-      },
-    },
-    {
-      fields: {
-        question: {
-          value: "Coldest planet in the solar system",
-        },
-        answer: "Duis ac magna vel urna pharetra iaculis vel et nibh. Mauris facilisis venenatis tempor. Sed porta, ipsum dictum rhoncus porttitor, neque odio sodales elit, at tincidunt eros lacus nec felis. Cras a porttitor mi, at ullamcorper neque. Aliquam at leo bibendum, ornare arcu sit amet, maximus ante. Fusce id enim.",
-      },
-    },
-  ];
 
   // state management for accordions:
   const [currentAccordion, setCurrentAccordion] = useState(null);
