+++ src/components/Allinone/Allinone.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useRef } from "react";
 import Layout from "../Layout/Layout";
 import Accordion from "../Accordion/Accordion";
 import CardComponent from "../CardComponent/CardComponent";
@@ -8,8 +8,10 @@ import Slider from "../Slider/Slider";
 import Table from "../Table/Table";
 
 const Allinone = () => {
+  const titlePage = "All at Glance";
+
   return (
-    <Layout>
+    <Layout pageName={titlePage}>
       <CardComponent />
       <Slider />
       <OrderList />
