--- src/components/Everything/Everything.js
@@ -1,22 +1,0 @@
-import React from "react";
-import Accordion from "../Accordion/Accordion";
-import CardComponent from "../CardComponent/CardComponent";
-import Footer from "../Footer/Footer";
-import Form from "../Form/Form";
-import OrderList from "../OrderList/OrderList";
-import Slider from "../Slider/Slider";
-import Layout from "../Layout/Layout";
-
-const Everything = () => {
-  return (
-    <Layout>
-      <CardComponent />
-      <Slider />
-      <OrderList />
-      <Accordion />
-      <Form />
-    </Layout>
-  );
-};
-
-export default Everything;
