+++ src/components/Table/Table.js
@@ -1,34 +1,6 @@
-function Table() {
-  const EMP_DATA = [
-    {
-      id: "1",
-      name: "Nirmal",
-      role: "front end developer",
-      project: "Boyd Loyalty",
-      occupied: "50%",
-    },
-    {
-      id: "2",
-      name: "shail",
-      role: "assistant lead tech",
-      project: "Bachmans",
-      occupied: "40%",
-    },
-    {
-      id: "3",
-      name: "Parth",
-      role: "front end developer",
-      project: "Mahindra",
-      occupied: "80%",
-    },
-    {
-      id: "4",
-      name: "Jay",
-      role: "associate front end developer",
-      project: "Seaworld",
-      occupied: "100%",
-    },
-  ];
+import EMP_DATA from "../../store/tableData";
+
+const Table = () => {
 
   return (
     <>
@@ -38,7 +10,7 @@ function Table() {
 
           {/* Correct Approach: */}
 
-          {/* <thead>
+          <thead>
                 <tr className="bg-cyan-400 text-2xl">
                     <th className="border border-slate-300">ID</th>
                     <th className="border border-slate-300">Name</th>
@@ -46,17 +18,17 @@ function Table() {
                     <th className="border border-slate-300">Project</th>
                     <th className="border border-slate-300">Allocation</th>
                 </tr>
-            </thead> */}
+            </thead>
 
           <tbody className="bg-cyan-100 text-lg">
             {/* Wrong Approach */}
-            <tr className="bg-cyan-400 text-2xl">
+            {/* <tr className="bg-cyan-400 text-2xl">
               <th>ID</th>
               <th>Name</th>
               <th>Role</th>
               <th>Project</th>
               <th>Allocation</th>
-            </tr>
+            </tr> */}
 
             {EMP_DATA.map((emp, index) => {
               return (
