+++ src/store/accordionData.js
@@ -1,0 +1,27 @@
+const DATA = [
+    {
+      fields: {
+        question: {
+          value: "Where is the sun located?",
+        },
+        answer: "Duis ac magna vel urna pharetra iaculis vel et nibh. Mauris facilisis venenatis tempor. Sed porta, ipsum dictum rhoncus porttitor, neque odio sodales elit, at tincidunt eros lacus nec felis. Cras a porttitor mi, at ullamcorper neque. Aliquam at leo bibendum, ornare arcu sit amet, maximus ante. Fusce id enim.",
+      },
+    },
+    {
+      fields: {
+        question: {
+          value: "Only human sustainable planet in the Solar System?",
+        },
+        answer: "Duis ac magna vel urna pharetra iaculis vel et nibh. Mauris facilisis venenatis tempor. Sed porta, ipsum dictum rhoncus porttitor, neque odio sodales elit, at tincidunt eros lacus nec felis. Cras a porttitor mi, at ullamcorper neque. Aliquam at leo bibendum, ornare arcu sit amet, maximus ante. Fusce id enim.",
+      },
+    },
+    {
+      fields: {
+        question: {
+          value: "Coldest planet in the solar system",
+        },
+        answer: "Duis ac magna vel urna pharetra iaculis vel et nibh. Mauris facilisis venenatis tempor. Sed porta, ipsum dictum rhoncus porttitor, neque odio sodales elit, at tincidunt eros lacus nec felis. Cras a porttitor mi, at ullamcorper neque. Aliquam at leo bibendum, ornare arcu sit amet, maximus ante. Fusce id enim.",
+      },
+    },
+];
+export default DATA;
