+++ src/store/menuList.js
@@ -1,0 +1,50 @@
+
+const menuItems = [
+    {
+        title: 'Tabs',
+        url: '/tabs',
+        submenu: [
+            {
+                title: 'Web design',
+                url: '#',
+            },
+            {
+                title: 'Web development',
+                url: '#',
+            },
+            {
+                title: 'SEO',
+                url: '#',
+            },
+        ],
+    },
+    {
+        title: 'Home',
+        url: '/',
+    },
+    {
+        title: 'Explore',
+        url: '/explore',
+    },
+    {
+        title: 'Gallery',
+        url: '/gallery',
+    },
+    {
+        title: 'FAQs',
+        url: '/faq',
+    },
+    {
+        title: 'Contact Us',
+        url: '/contact',
+    },
+    {
+        title: 'Data',
+        url: '/table',
+    },
+    {
+        title: 'All at Glance',
+        url: '/all-at-glance',
+    }
+];
+export default menuItems;
