+++ src/store/sliderData.js
@@ -1,0 +1,27 @@
+const SLIDES_DATA = [
+    {
+      id: "1",
+      title: "slide 1",
+      src: "https://wallpaperaccess.com/full/441330.jpg",
+      alt: "sd_1",
+    },
+    {
+      id: "2",
+      title: "slide 2",
+      src: "https://www.nasa.gov/sites/default/files/styles/ubernode_alt_horiz/public/thumbnails/image/transp_img0056_small.jpg",
+      alt: "sd_2",
+    },
+    {
+      id: "3",
+      title: "slide 3",
+      src: "https://www.nasa.gov/sites/default/files/styles/ubernode_alt_horiz/public/thumbnails/image/70hc958_banner.jpg",
+      alt: "sd_3",
+    },
+    {
+      id: "4",
+      title: "slide 4",
+      src: "https://www.nasa.gov/sites/default/files/styles/ubernode_alt_horiz/public/thumbnails/image/1es_agn_aurore.jpg",
+      alt: "sd_4",
+    },
+];
+export default SLIDES_DATA;
