+++ src/store/tableData.js
@@ -1,0 +1,31 @@
+const EMP_DATA = [
+    {
+      id: "1",
+      name: "Nirmal",
+      role: "front end developer",
+      project: "Boyd Loyalty",
+      occupied: "50%",
+    },
+    {
+      id: "2",
+      name: "shail",
+      role: "assistant lead tech",
+      project: "Bachmans",
+      occupied: "40%",
+    },
+    {
+      id: "3",
+      name: "Parth",
+      role: "front end developer",
+      project: "Mahindra",
+      occupied: "80%",
+    },
+    {
+      id: "4",
+      name: "Jay",
+      role: "associate front end developer",
+      project: "Seaworld",
+      occupied: "100%",
+    },
+];
+export default EMP_DATA;
